cover.sh: don't copy coverage files
cover.sh was previously copying each of the .gcda and .gcno files
to a directory, squashing the original directory hierarchy. This
resulted in some of the coverage data files being lost, eg both
userclk and coreidle have a main.c file. The change captures the
coverage data from the files in the directory hierarchy where they
are generated.