add test_pyopae binary and some tests (#812)
* add test_pyopae binary and some tests

test_pyopae binary uses pybind11 to embed the python interpreter and to
create opae.fpga as an embedded module along with test_system,
test_platform, and test_device.

test_pyopae uses the first argument to determine if it should run unit
tests defined in an external file. If the first command line argument is
`test`, it evaluates the script identified by the next argument.
This script is assumed to contain tests compatible with the Python
unittest module.

In order to embed the opae.fpga module, a preprocessor definition was
added to control how the pybind11 module is defnined (embedded or not).
It also exposes fpgaInitailize as a module method, `initialize` when it
is embedded. This allows calling initialize when running on mock
devices.

This commit also adds two Python test scripts, test_properties.py and
test_shared_buffers.py and adds CMake tests for running them with
test_pyopae.

test_system is changed to include an atomic bool variable to indicate if
it has been initialized. If it hasn't, it will pass through to glibc
calls open and close.
TODO: Add similar logic to the other mocked calls.

* add pylint directives to ignore warnings

Because these test scripts are designed to run with the embedded
interpreter, pylint will complain about variables that are imported
using pybind11 C++ import functions.