Fixes to support different FME/Port numbers
In certain device topologies, it has become valid to have
an FME/Port device number that differs from the
containing device number, eg

/sys/class/fpga/intel-fpga-dev.1/intel-fpga-fme.0
/sys/class/fpga/intel-fpga-dev.1/intel-fpga-port.0

This fix adds a subdev_instance member to struct _fpga_token and
tracks both the device_instance (1 in the example) and the
subdev_instance (0 in the example) in the struct. 5 of the sysfs
accessor functions are also updated to accomodate both device and
subdev numbers.