pydiag: fix setup method to include `in_args`
The base version of diagtest.setup includes `in_args` but nlb3 does not.
The purpose of this is to allow for passing in command line arguments
(instead of relying on sys.argv vector). This fixes the call to
parse_known_args to use it and also adds the argument to nlb3.setup
method.