update fpga/__init__.py for sysobject
Include sysobject in the opae/fpga/__init__.py namespace package file.
Without this, the sysobject class will not be visible in the opae.fpga
namespace.
While it is not necessary for it to exist in the namespace (as
sysobject instances can only be created from tokens/handles), it is
necessary for it to be in the namespace in order to get to its
documentation.