fpgad: bring fpgad to date with API changes
Update fpgad to enumerate for supported devices, rather than assuming
that it is running on a dual-socket MCP. This required changing the
event registration mechanism to move away from using the sysfs file as
the device identifier. The fme/port device's object ID is used instead.
Update test cases to comply with the new design.