add sysfs_initializion routine
the sysfs_initialize and sysfs_finalize is meant to be called when the
xfpga plugin is loaded so that it will discover devices (regions) and
device resources (fme or port). This uses a format table for looking for
sysfs patterns used by the upstreamed driver or by the intel driver.
Once regions/resources are discovered, they are placed in an array of
regions.

TODO:
Add accessor functions to get regions/resources from this internal data
structure.
Modify fpgaEnumerate to use these new sysfs accessor functions.