xfpga: switch to using sysfs ptr
remove use of static `_sysfs_format_index` variable
replace with `_sysfs_format_ptr`

Because the `_sysfs_format_index` is initialized to zero, it is hard to
determine if a driver has been discovered. Changing it to a pointer that
points to a sysfs_formats entry can instead be used to indicate whether
or not a driver has been discovered.