opae-c: add dummy plugin and basic test
This dummy plugin makes use of the configuration string by parsing it
and then using the values. The first two values are used to print out to
stdout which the tests verifies. The third value is the number of dummy
tokens to create which is then used for the rest of the test for
enumerating and opening/closing tokens.

This also fixes the ASSERT_NOT_NULL_MSG_RESULT macro to include a space
before "is NULL" in the mssage.