opae-c: add dummy plugin and basic test (#1117)
* opae-c: add dummy plugin and basic test

This dummy plugin makes use of the configuration string by parsing it
and then using the values. The first two values are used to print out to
stdout which the tests verifies. The third value is the number of dummy
tokens to create which is then used for the rest of the test for
enumerating and opening/closing tokens.

This also fixes the ASSERT_NOT_NULL_MSG_RESULT macro to include a space
before "is NULL" in the mssage.

* fix plugin env. in CMake

Set LD_LIBRARY_PATH in env. for pluginmgr test.
Also, redirect stderr to /dev/null to limit the output generated in
CI.

* add debug printing on failure

Print out LD_LIBRARY_PATH env. var if loading plugins fails.

* add dependency for plugin tests on dummy_plugin