Rrojo/plugin mgr cfg (#1143)
* Fix issues identified by checkers.

* Update platform RTL Quartus IP from 18.1. (#1116)

- The only synthesis-visible change is a new timing constraint on Avalon DC FIFO.
- Update the script the pulls sources from a Quartus release due to file name changes.

* Fixed leak in xfpga sysfs tests. (#1118)

* Retry socket select operation if it fails with EINTR (#1115)

reference:
https://www.gnu.org/software/libc/manual/html_node/Interrupted-Primitives.html

* opae-c: don't apply versions to modules

Attempting to apply VERSION/SOVERSION to module libraries is causing
build errors on distros with newer versions of cmake. This change
removes the code that was doing that, and it changes the library type of
libbmc.so from SHARED to MODULE (it is dynamically loaded).

* remove var. name in catch statement (#1120) (#1121)

When catching 'not_found', simply return 0.
No need for processing exception variable.

Signed-off-by: Rodrigo Rojo <rodrigo.rojo@intel.com>

* Fixed leak in sysobject api sample. (#1126)

* Add RPM specfile under the ownership of ci team (#1107) (#1125)

* Nkorde/cherry pick from 1.3.0 (#1129)

* Update release to 1.3.0-2 (#1105)
* Default install location to  /usr only for package generation

* Cherry pick new fpgad implementation

Move fpgad to a plugin-driven design. See monitored_device.h for details.

* xfpga: use modbmc for metrics API

Build libmodbmc.so as a shared module library for use with xfpga's
metrics API, because the API wants to dlopen the library. Leave
libbmc.so as a shared library object for use with pacd.

* xfpga: fix for cstr_dup

Separate the malloc check from the copy check so that a successful
allocation can be freed on copy error.

* update packaging for bmc and libbitstream (#1134)

* Add bmc libs to opae-libs deb and rpm package

* Add versioning back to libbmc.so

* Remove unnecessary content from tools-extra RPM package(#1136)

* testing: add mock_opae class (#1138)

`mock_opae_p` class is a templated class that derives from TestWithParam
and is meant to be a mock class for mock platforms.
The template argument is used to specify how many tokens to use
(if any).

This is a first step in refactoring tests to avoid code cloning.

* Fix compilation of fpgad on older systems. (#1140)