merge feature/plugin-mgr-config into master (#1147)
The major changes in this is the ability to load plugins from a plugin configuration file.
It looks for configuration files in the following locations:
"/etc/opae/opae.cfg",
"$HOME/.config/opae/opae.cfg",
"$HOME/.local/opae/opae.cfg",
"$HOME/.local/opae.cfg"

And will attempt to load plugins from the first one found.
This also adds unit tests, including a dummy plugin, that can be used as reference also when creating a plugin.

* opae-c: add dummy plugin and basic test

This dummy plugin makes use of the configuration string by parsing it
and then using the values. The first two values are used to print out to
stdout which the tests verifies. The third value is the number of dummy
tokens to create which is then used for the rest of the test for
enumerating and opening/closing tokens.

This also fixes the ASSERT_NOT_NULL_MSG_RESULT macro to include a space
before "is NULL" in the mssage.