fpgaflash: use BMC RSU (reboot) for FPGA RSU
Change the FPGA RSU from using PFL based reconfiguration
to using the BMC RSU which essentially reboots the board.
This change works around the problem of the FPGA RSU
not being able to recover from from a failed update without
a power cycle.  By using the BMC RSU flow, if the FPGA
user image is corrupted, the bootflow will failover to
loading the factory image as expected.

Signed-off-by: Matthew Gerlach <matthew.gerlach@linux.intel.com>