remove 'resolve_file_name' function in pluginmgr
Switch to a different way of getting a user's home directory that does
not depend on the $HOME env. variable.
This way uses getpwuid and getuid to get the passwd structur that
includes the user's home director.
In 'find_cfg', swith to iterating over two sets of possible config
paths. The first set lists possible paths relative to a user's home
directory.