pluginmgr: fix bugs in finding cfg file (#1154)
* pluginmgr: fix bugs in finding cfg file

* reorder candidate config files so that user config files are found
first
* add /usr/local/etc/opae/opae.cfg as a candidate config file
* fix but in resolve_file_name when HOME env. variable is longer than
input string

* remove 'resolve_file_name' function in pluginmgr

Switch to a different way of getting a user's home directory that does
not depend on the $HOME env. variable.
This way uses getpwuid and getuid to get the passwd structur that
includes the user's home director.
In 'find_cfg', swith to iterating over two sets of possible config
paths. The first set lists possible paths relative to a user's home
directory.

* Add parameterized test for finding and parsing a config file under a user's home directory