xfpga: fix VF enumeration bug
The current enumeration flow assumes that both fme and port nodes in
sysfs have the same parent (device) node. With this assumption, some
code related to either finding an fme node from a port node, or vice
versa, looks for the target node by looking in the current node's parent
directory.

For example, say you have a device with the following fme and port
nodes (using conceptual names instead of names specific to a specific
version of the OPAE kernel driver):
```
/sys/class/fpga/dev.0/fme.0
/sys/class/fpga/dev.0/port.0
```
Given a path to an fme node, the port node can be found by the following
glob pattern:
`/sys/class/fpga/dev.0/fme.0/../*port.*`

or looking forn an fme node using the path to the port node:
`/sy/class/fpga/dev.0/port.0/../*fme.*`

The paths change, however, when SR-IOV is enabled (a port device is
released) and a VF is created. The steps to do this would look something
like (again omitting driver specific node names like `intel-fpga`):
```bash
fpgaport release /dev/fme.0 0
echo 1 > /sys/class/fpga/dev.0/device/sriov_numvfs
```
The result sysfs paths for the fme and port nodes woule now look
something like:
```
/sys/class/fpga/dev.0/fme.0
/sys/class/fpga/dev.1/port.1
```
This change fixes the assumption mentioned above and uses symlinks
(physfn or virtfn?) present when a port has been migrated to a VF.

Now, when looking for a port node given a path to an fme node, the code
looks at the following glob patterns:

```
/sys/class/fpga/dev.0/fme.0/../device/virtfn*/fpga*/dev.*/*port.*
/sys/class/fpga/dev.0/fme.0/../device/fpga*/dev.*/*port.*
```
_Note_ Only one of the preceding glob patterns will exist at a time
depending on whether or not a port is on a VF (the first pattern) or on
a PF (the second pattern).

Similarly, when looking for an fme node given a path to a port node, the
following patterns will be tried:

```
/sys/class/fpga/dev.0/port.0/../device/physfn/fpga*/dev.*/*fme.*
/sys/class/fpga/dev.0/port.0/../device/fpga*/dev.*/*fme.*
```

_Note_: Some changes to sysfs.c may not seem related to this fix.
That is because the original plan was to use regular expressions
for finding the nodes for "dev.*', "fme.*", or "port.*". Part of this
platn required that the regex pattern matching routines be refactored in
ordre to reuse them from other functions.
This idea was abandoned, however, in favor of using simpler glob patterns
(which doesn't require opening a directory and iterating over its
contents).

Finally, for testing, a mock sysfs file structure was added to to
represent a scenario when a port is on a VF. The `test_device` class and
the mock platform db had to be modified to include the number of VFs in
a given device. Several tests had to be modified to use this information
in some of their assertions.