fix bug in re_match_(device|region)
This fixes a bug in re_match_device and re_match_region where it was
using (ptr-end) when calculating at what point in the string to place a
null byte in to terminate the string.

This was resulting in a negative number and the pointer being
dereferenced pointed to somewhere else in the stack of the calling
function thus corrupting it.

The fix is to swap the pointers to correctly calculate the size of the
string (end-ptr).