fix bugs related to use of `realpath`
Recent changes use `realpath` to get the realpath (without symlinks) to
sysfs nodes and uses that for comparing/identifying certain kinds of
nodes in the OPAE sysfs tree. According to the man page of realpath (3),
there are bugs and it states that "it is impossible to determine a
suitable size for the output buffer".

This change assures that the buffers that are ultimately used as the
output variable to the call to realpath are of size PATH_MAX
(4096 as defined limits.h).

It has also been observed that calling realpath with a buffer smaller
than PATH_MAX can overflow the buffer even though the lengths of the
resolved paths are ~128.