fix samples and add a note to doxygen of c_type fn
Fix the C++ sample for hello_fpga-1.cpp to use explicit types instead of
the `auto` keyword. This is done to make it clearer as to the types of
objects being created.

Add a note in shared_buffer::c_type doxygen markup to warn developers of
the dangers of _stealing_ a raw pointer from the shared_buffer object.