merge super-rsu [phase 1] (#708)
* add skeleton of fpgaboard script (#685)

* add skeleton of fpgaboard script

This prototype of fpgaboard is responsible for finding a PAC board given
a PCI address ([segment:]bus:device.function).

Once it finds the board, it can either
* print out the current versions of programmable parts
 (user image, bmc_fw, max10)
* update the image by calling fpgaflash with the correct commands
  assembled by processing the configuration file (in JSON format).

A sample config file, rsu.json, is checked in for reference to the
schema which is still WIP.

Currently, this script does not actually spawn an fpgaflash process but
instead can just print out the command that would be invoked.

This serves as a starting point for the "super rsu" flow and implements
checking versions of bmc_fw, max10, and the a10 user image.

* change logging to LOG in fpgaboard, update JSON (#687)

* Change logging in fpgaboard to use a named logger, not the root logger.
* Update the sample JSON schema to include a key for the product name.
  This product name will be used to map to vendor/device ids

* Add skeleton of super-rsu script (#689)

* Add skeleton of super-rsu script

This currently discovers pci devices identified by the product in the
RSU config file.
For every device discovered, it will call 'fpgaboard' script to update
the board using the flash images listed in the config file
This also stubs out the logic for performing the RSU at the BMC level.


* add aer_disabled context manager
Disabling AER for a given device actually disables AER for its parent.
Use the context manager with the parent nodes

* Add sysfs_node class as the base sysfs object
sysfs_node class will provide the methods to read/write attributes
relative to the node's root (as indicated by the property sysfs_path)

* Add pci_node to inherit from sysfs_node
* Add fpga_node class that is meant to represent the fpga node under a
pci_node tree. The fpga_node will be expected to have nodes to represet
fme and port objects.

* move vendor/device id info to rsu config (#692)

* filter out child nodes that begin with pci addr

When finding children of a pci node, filter out subdirectories that
begin when the same pci address as the current node.

* remove internal PRODUCT table

move product vendor/device id into rsu.json file

* change copyright date to 2019

* fix type in call to `startswith`

the filter (for omitting items that start with the pci address of the node) in _find_children
was 'startwith' instead of 'startswith'

*  add support for using 'additional_devices' in cfg (#693)

* add support for using 'additional_devices' in cfg

create a new class, board_info, that contains:
* the pci_node object
* the fpga_node object
* _other_pci objects
  _other_pci is assigned from the 'additional_devices' object in the
  JSON config. If set, this will be used to find the common root port
  of both the pci_node of the board_info, and all additional devices
  identified by the set of (vendor, device) tuple

create a helper function, 'find_subdevices', that will return a set of
all devices below the current node, as identified by (vendor, device)
tuple. This helper function is used when finding the common root of the
pci_node in board_info and any additional devices.

* fix bug in context manager, 'aer_disabled'

The context manager was yielding the last node in the list of nodes
which is a problem especially if no nodes are passed in.
Instead, it is yielding True if the list of nodes is not empty and if
all of the nodes are non-null (not None)

* updates to super-rsu (#702)

* updates to super-rsu

This commit has lot of updates of super-rsu that make it nearly
functionally complete.

Several components from 'fpgaboard' have be incorporated to enable
flashing. Several components/routines have been refactored.

At a high level this can now:
* parallelize multiple flash routines
  * process the rsu config and make a list of tasks for flashing
    * images will be flashed only if versions are not equal or
      force-flash is indicated on the command line
  * for every board discovered, spawn a thread that runs the tasks
  * handle timeout by terminating threads/tasks
    * TODO: handle factory flash timeout differently
* perform rsu
* remove/rescan of the first common ancestor of the main device and
  any additional devices (specified in rsu config)
* log 'trace' messages (very verbose)

* support 'force' field in rsu json

Change the default, action, and destination in the --force-flash arg to
be more intuitive

Look for 'force' in the task config for a rsu flash item and use this to
force flashing of an image

* remove fpgabaord, add super-rsu to install