fpgaflash: abort if pacd, fpgad, or fpgainfo are running
fpgaflash takes a big performance hit if the sensors are
being polled during the flash. Also, during a NIOS FW update,
the sensor values are unchanging and therefore meaningless.
This is a short-term solution to mitigate the effects or
running polling the MAX10 register during flash by instructing
the user to shut down these tools before running fpgaflash.

Signed-off-by: Russ Weight <russell.h.weight@intel.com>