super-rsu: fix rsu routine for aer/remove/rescan (#763)
Previously, the rsu routine involved finding the first common ancestor
of the target device (0b30) and the additional device (0b32) and using
for disabling aer and rescanning the buses under this device.

If the additional device is incorrect in the manifest, looking for the
common ancestor will result in finding just the parent of the target
device. This will result in disabling AER on its parent (not the root
port) which can cause the system to reboot when RSU is triggered on the
FPGA

This fix uses the root device with respect to the target device (0b32)
and removes the dependency on the "additional device".