xfpga: revert difference in BBS version
A recent change introduced a difference in the bitfield encoding
for the bitstream_id CSR for MCP vs. DCP. It has been discovered
that this difference was a bug in the CSR spreadsheets for DCP.
With this change, both MCP and DCP assume the same encoding for
the bitstream_id and determine the version in the same way.