fpgad-vc: sensor overrides in config
* fpgad-vc: sensor overrides in config

Allow users to override the sensor trip points using the configuration
file settings.

To do so, add..

"config-sensors-enabled": true,
"sensors": [
	{
		"id": 1,
		"high-fatal": 100.0,
		"high-warn": 90.0,
		"low-warn": 10.0,
		"low-fatal": 0.0
	},
	...
]

If config-sensors-enabled is false, then the entire sensors section
is ignored. If id is not a valid sensor id, then the entry has no
effect. Each of high-fatal, high-warn, low-fatal, and low-warn are
optional. If omitted, then the sensor value is not checked for the
corresponding limit.

* Implement cap of user limits by hw when applicable