opae: merge master changes to feature branch (#1295)
* fpgad-vc: graceful shutdown models super-rsu flow

Update the method used to remove the device driver to that of super-rsu:
* Write 1 to the device's reset node.
* Write 1 to the remove node of the root bridge's child. (removes
all VC devices)

When rescanning,
* Write "on" to the power/control node of the root bridge's child.
* Write 1 to the rescan node of the root bridge's child.

* fpgad-vc: don't write device reset node

Remove the code that was writing 1 to the root bridge device's
reset sysfsnode.
Fix the tripped sensor counter so that it correctly measures for
each device.
Remove extraneous cool down lock.

* fpgad-vc: sensor overrides in config

* fpgad-vc: sensor overrides in config

Allow users to override the sensor trip points using the configuration
file settings.

To do so, add..

"config-sensors-enabled": true,
"sensors": [
	{
		"id": 1,
		"high-fatal": 100.0,
		"high-warn": 90.0,
		"low-warn": 10.0,
		"low-fatal": 0.0
	},
	...
]

If config-sensors-enabled is false, then the entire sensors section
is ignored. If id is not a valid sensor id, then the entry has no
effect. Each of high-fatal, high-warn, low-fatal, and low-warn are
optional. If omitted, then the sensor value is not checked for the
corresponding limit.

* Implement cap of user limits by hw when applicable
(cherry picked from commit 2c8c1e87f1257fe65df7bf1084a48921e348c7e5)

* Add fpgaflash to deb tools-extra package (#697) (#1278)

* Added missing tools to the tools-extra deb package. (#1281)

* Fix a few compilation problems on old compilers. (#1285)

* ASE: Add ASE as plugin (#1265)

* Fixed the plugin config read size to account for a null terminator. (#1290)

* tools: add support board plugin modules to fpgainfo (#1239)

* tools: add support board plugin modules to fpgainfo

   - Dynamically load plugin module
   - Support MAC, PHY group commands
   - Fix Invalid Metrics value

* xfpga: fix impl. of ObjectGetSize (#1293)

Syncing an object is only applicable to objects that are backed by sysfs
files (and not directories or containers). This fix ensures that
sync_object is called only for file objects.

* api: implement missing SysObject API functions (#1294)

* api: implement GetObjectAt API function in opae-c

* xpfga: add GetObjectAt to plugin adapter

* api: add ObjectGetType API to api and xfpga

* api: fix fpgaCloneObject
** copy the object type
** allocate the buffer object for file objects