opae.admin: initial check-in of opae.admin pkg
opae.admin is meant to be a Python package for administrative tasks and
uses the interfaces exported by the OPAE kernel driver.
This first check-in supports basic discovery of sysfs nodes and creates
objects that are instances of sysfs_class class. Client applications can
extend this by deriving from the sysfs_class.