Improve performance of ASE fpgaPrepareBuffer checker.
The preallocated page checker confirms two things: that a page is mapped
and that the size is what the user claims. It is attempting to mimic the
behavior of the kernel driver. Checking is pretty slow, which is generally
OK when using huge pages since there aren't very many. When using 4KB pages
it can be painful. The new code does two things to optimize the 4KB case:
 - Use /proc/self/maps instead of /proc/self/smaps, since only the address
   range has to be checked and not the page size.
 - Don't check every page. The cost is too high for the value. The first
   200 pages are checked, which ought to catch the majority of user code
   bugs. After that, every 100th new page is checked. I thought about
   testing randomly, but having the result be unpredictable seems like
   a bad idea.