super-rsu: add support for fpgasupdate (#895)
* super-rsu: add support for fpgasupdate

All manifests versioned 4 and higher will now have to include secure
images. The version will be used to indicate that the manifest is
made of secure images.

This also introduces a new "type" value:
* bmc_pkg

It is handled by the new flashable class, bmc_pkg, which is composed of
bmc_img and bmc_fw. T

It will be expected for the manifest to include a list for the
versions.

The revision can remain an one item which applies to both.

```JSON
  "version": ["a.b.c", "a.b.d"],
```