super-rsu: fix template for nvmupdate cfg  (#908)
* super-rsu: fix templates for nvmupdate cfg

Add VENDOR/DEVICE to the template string used when generating an
nvmupdate config file. These will get replaced from the data in the
nvmupdate section of the manifest.

* super-rsu: update 'Family' in nvmcfg template

Although the 'Family' GUID in the nvmupdate config file template is a
comment, this udpates it to the latest GUID from the latest releases of
FVL firmware and nvmupdate.