super-rsu: make super-rsu a bit smarter (add find_config) (#26)
* super-rsu: make super-rsu a bit smarter

* Add find_config method that looks for one suitable rsu config file
  * This looks in /usr/share/opae and all subdirectories for *.json
    files
  * For every *.json file found, it checks:
    * Does it include "program": "super-rsu" key/value
    * Can it enumerate more than one FPGA with this vendor/device
    * If the above two criteria are met, it adds it to a list of
      'candidates'
  * If candidates is exactly one, then there should be no ambiguity so
    it returns an open file object with this file
* Change the rsu_config argument to optional
* If the rsu_config is absent, set it with a call to find_config

This makes it so that super-rsu can be called with no command line
arguments in the simplest case (a system with only one type of card
installed)

* super-rsu: add 'program' command line argument.

'program' command line argument is used when looking for a super-rsu
manifest file. If the 'program' value is found in the manifest, it will
use this manifest.

This makes it so that two super-rsu manifests can live together but be
differentiated by their 'program' value. This can be useful for the otsu
requirement on vista-creek that a temporary max10 be flashed before
calling fpgaotsu.

This way, one can do something like:

```shell
super-rsu --program enable-otsu
fpgaotsu
super-rsu
```

Here, 'enable-flash' will be the value in the manifest that loads the
temporary max10 image that opens up the secondary flash device needed by
fpgaotsu.

```JSON
{
  'program': 'enable-otsu',
  'product': 'n3000',
  'vendor': '0x8086',
  'device': '0x0b30',
  'flash': [
    {
      'type': 'bmc_img',
      'force': true,
      'filename': 'temp-max10.rpd'
    }
  ]
}
```

* super-rsu: handle different schema in find_config

If the find_config routine is inspecting a json file, its scema may be
different that that which is compatible with super-rsu.
This handles the case when the data parsed is not even a dictionary
item.

* super-rsu: fix find_config to cast vendor/device

vendor/device are stored as strings in the maniefst but the enumeration
expects them as integers.

* super-rsu: find_config: add debug log message

Add debug message when considering manifest files.