fpgaotsu_09C4: improve the look and feel
Updates the output to follow the model of fpgasupdate - if log
level is < INFO, ie if it is DEBUG, then send the progress object
output to LOG.debug, producing one line of output for each %
increase in progress. If log level >= INFO, then send the progress
object output to stdout so that the entire progress consumes only
one line.