fpgasupdate: calculate write_block size
Dynamically adjust the block size for fw_write_block() so that
the duration of the fw_write_block() call is as close to a given
target duration as possible. This guarantees that the user will
see progress meter updates at nearly fixed intervals. The
measurements are of "wall time", so they will be agnostic to the
hardware platform.

* Add log output for CHECK_COMPLETE ioctl