ASE: Ignore CCI-P c1Tx sop when not valid (#1366)
A nasty bug in ASE emulation of CCI-P c1Tx: when sop was set in a c1Tx
header the mdata and address were latched, even when the valid bit is
not set. Even worse, the emulator's debug logging printed correct
behavior despite the internal emulation using the wrong values.