fpgasupdate: don't prevent running with fpgainfo
The check for fpgainfo has proven too restrictive, because it
only checks for the existence of the process (and doesn't know
which device that process is actually targeting.) This has proven
problematic in multi-card scenarios, where a customer may want to
run fpgainfo targeting one card while performing an update on
another.

Remove this check for the update apps.