libopae: fix error API clear routine
The current method to clear errors reads the value of the error and
writes it back to the clear file. Clearing ras injected errors is
different in that it must be a zero to clear (instead of the current
value). This adds a check for the error name being cleared and uses zero
if it is 'inject_error'.

Signed-off-by: Rodrigo Rojo <rodrigo.rojo@intel.com>