Add reference counting to ASE's pinned buffer model. (#1440)
When pinning pages on real FPGAs, the kernel keeps a reference count of
the number of times a page is pinned. Pages must be released the same number
of times before the kernel unpins them. ASE was not keeping the equivalent
count, causing translations to be lost the first time a buffer was released.
New code now uses 8 bit counters to track references.