Add cmake option BUILD_TOOLS_EXTRA
This option controls if the applications in tools/extra are built.
The default is 'ON'

Some tools depend on hwloc.  When the hwloc version changes,
it is necessary to disable building extra/tools while the updating
the codebase.

This works around a problem with hwloc 2.0 error when building on
fedora 31

.../tools/extra/pac/fpgabist/dma_vc/fpga_dma_test.c:707:7: error: 'HWLOC_TOPOLOGY_FLAG_IO_DEVICES' undeclared (first use in this function); did you mean 'HWLOC_TOPOLOGY_FLAG_IS_THISSYSTEM'?
  707 |       HWLOC_TOPOLOGY_FLAG_IO_DEVICES);
      |       ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
      |       HWLOC_TOPOLOGY_FLAG_IS_THISSYSTEM