Trix/master/sa
* Rename element 'type' to 'sensor_type'

For consistency, the element 'type' also called 'sensor_type'.
Use sensor_type because it is more descriptive.

* Consolidate sensor type enums to BMC_SENSOR_TYPE

In the latest compilers, both gcc and clang there is this error

.../libopae/plugins/xfpga/metrics/bmc/bmc.c: In function ‘bmcGetSDRDetails’:
.../libopae/plugins/xfpga/metrics/bmc/bmc.c:475:16: error: implicit conversion from ‘BMC_TYPE’ to ‘BMC_SENSOR_TYPE’ [-Werror=enum-conversion]
  475 |  details->type = this_val->sensor_type;
      |                ^
cc1: all warnings being treated as errors

To fix this problem, use a common enum type.
BMC_SENSOR_TYPE was picked because it appears to be the public interface.

Co-authored-by: Ananda Ravuri <33236856+anandaravuri@users.noreply.github.com>