buffers: fix 2M huge page allocation
The previous flags for allocations in the 2MB range were broken,
allowing 2MB buffer allocations to be satisfied from the 1GB pool.
Restrict the 2MB allocations to the 2MB pool. This also corrects
a munmap() error when passing an incorrect buffer length (2MB was
passed when the allocations were actually 1GB).