xfpga: correct the enumeration code / fix for containers (#49)
The enumeration code was creating the candidate device list in the wrong order. Aside from difficulty debugging it (due to the incorrect list order), it was also performing an unnecessary fetch of the socket id from the FME when enumerating the AFU.

This change corrects the enumeration code so that the candidate device list is ordered per the original intent. It also separates the creation of the candidate device list from the "sync'ing" of the resulting FME/Port properties. Before "sync'ing" the properties, a candidate device's sysfspath is verified to exist as a directory; and a candidate device's devpath is verified to exist as a character device. This is done so that, in environments where the FME may not exist, no token is returned for the FME device.