xfpga: fix irq registration
event.c in xfpga is always using 0 for the irq num when reqistering for
events or unregistering for events.
This fix changes it so that the xfpga API implementation keeps track of
this in both the internal data structures for handle and event_handle.
The handle data structure stores the number of irqs supported and this
is used when checking if this number has already been set or not.
The event_handle data structure stores the irq number and uses this when
unregistering the event.