xfpga: fix irq use in events API (#71)
The events API was always using index 0 when registering for interrupts.
Use the number of interrupts available along with a bitset of interrupts
for keeping track of interrupts when assigned/unassigned.
Also, save the interrupt index in the event_handle data structure for it
with the event. This is used when unregistering for events.