tests: add test_opae.py
This is a pytest file to run google test binaries.
This is intended for CI and expected to run in the build directory.
It will look in bin for test_* files and execute them
(with the exception of test_pyopae).
For test_pyopae, it looks in the build directory for test_*.py and runs
those with test_pyopae.