mmlink: update mmlink to support new IP
This updates mmlink to support both legacy remote debug interface
protocol as well the latest remote debug interface IP.

This will be the beginning of refactoring mmlink to be more modular and
easier to maintain.

This current change:
* Adds an interface class, remote_dbg
* Wraps the core mmlink logic into a legacy_dbg impl. class, legacy_dbg
* Wraps the new logic to support the new streaming interface in the
  remote debug IP.
* The current main routine instantiates the legacy_dbg class.

TODO: Use the DFH revision to switch between legacy_dbg or stream_dbg
classes.