Split tarball generation for rpm generation
The offical fedora rpm depends on the url source tarball.
Split tarball generation out so it is easier to generate.

Generalize the naming of the tarball to follow the convention
of <package>-<version>.tar.gz.

Fix opae.spec to use the new source loction opae-<version>/

Generalize the tarball excludes all hidden files

Signed-off-by: Tom Rix <trix@redhat.com>