update fedora spec file and tarball generation script
The offical fedora rpm depends on the url source tarball.
Split tarball generation out so it is easier to generate.

Generalize the naming of the tarball to follow the convention
of -.tar.gz.

Fix opae.spec to use the new source loction opae-/

Generalize the tarball excludes all hidden files

Signed-off-by: anandaravuri <ananda.ravuri@intel.com>