vfio-plugin: support vf_token
Opening a VF bound to vfio-pci requires a secret token (in uuid) format
shared with the PF. This change adds support for this by detecting if
the device being opened is has a symlink to a PF (phyfn).
If so, it will open the PF with a random UUID and use the same UUID with
the VF being opened.

* create a new struct, vfio_pair_t
  * contains uuid struct and two opae_vfio ptrs
* create helper functions
  * read_pci_link
    * read a symlink in a sysfs pci device path
  * open_vfio_pair
    * allocate memory for vfio_pair_t
    * check for a physfn symlink
    * if found, generate a UUID and open both the physfn and the device
      using that token.
  * close_vfio_pair
    * call vfio_close on the opae_vfio ptrs (if non-null)
    * free memory allocated with open_vfio_pair