Rrojo/pci device updates (#1822)
Several improvements as well as sub-commands to pci_device script.

Change the first positional argument to be either the pci address or a
vendor/device ID pair. In either case, this input is treated like a
filter for enumerating devices and whatever sub-command is given will
operate on the resulting list of devices returned from this enumeration.

Add aer subcommand

This subcommand has it's own subcommands: dump or mask (with default
being dump)
dump subcommand will print out the aer error counters as reported
by the sysfs files for the device.
mask can either print out the current AER mask bits or set them
If show or print (or nothing) is given after the mask
command, it will show the current mask bits for AER
If all is given after the mask command, it will mask all bits
(by setting the values to 0xffffffff and 0xffffffff).
If off is given after the mask command, it will unmask all
bits ( by setting the values to 0x0 and 0x0).
If two numbers are present after the mask command, those two
numbers will be used to set the mask bits.
Update documentation to reflect these changes.