test_fpgaconf_c: updates for argsfilter
Now that we're using argsfilter, the input argv array can be
mutated when options are matched. To fix this in tests that call
fpgaconf_main multiple times, we need to reset the argv array
before every call.

argv must also be null-terminated per the standard.
(5.1.2.2.1 Program startup)
"argv[argc] shall be a null pointer."