cmake: refactor use of spdlog (#1863)
* cmake: refactor use of spdlog

The primary objective in this change is to allow for using a
preinstalled version of spdlog or fetch it as an external if not found.
Summary of changes/additions:
* add cmake/modules directory (move findsphinx to cmake/modules)
* add Findspdlog.cmake that tries to find spdlog
* call find_package(spdlog...) in external/CmakeLists.txt
  * if found, set spdlog_defines compile definitions to
    SPDLOG_COMPILED_LIB
  * else if not found, fetch it using opae_external_project_add but
    also force setting cmake variables that would be set if it was found
    * set spdlog_INCLUDE_DIRS to path under external subtree
* Change afu-test/CMakeLists to use:
  * ${spdlog_INCLUDE_DIRS}
  * ${spdlog_LIBRARIES)
  $ ${spdlog_DEFINES}

* cmake: fix wording/spelling in description

* cmake: add spdlog_DEFINITIONS variable

cmake find_package conventions uses <package>_DEFINITIONS.
Because of this, this change adds it in Findspdlog.cmake and changes its
use in afu-test/CMakeLists.txt.
This also sets spdlog_FOUND to true after adding it as an external.
This prevents it from trying to find it again.

* cmake: set spdlog version to 1.8